Fix bug and unit test failure in HHWheelTimer
Summary: Some minor bug and a failed test caused by D1578466

Test Plan:
fbconfig thrift/lib/cpp/test:HHWheelTimerTest && fbmake
runtests

Reviewed By: davejwatson@fb.com

Subscribers: trunkagent, alandau, bmatheny, njormrod

FB internal diff: D1581949

+++ folly/io/async/HHWheelTimer.h
@@ -219,7 +219,7 @@ class HHWheelTimer : protected folly::AsyncTimeout,
   typedef Callback::List CallbackList;
   CallbackList buckets_[WHEEL_BUCKETS][WHEEL_SIZE];
 
-  uint32_t timeToWheelTicks(std::chrono::milliseconds t) {
+  int64_t timeToWheelTicks(std::chrono::milliseconds t) {
     return t.count() / interval_.count();
   }
